backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
django-environ==0.11.2
PyJWT==2.8.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
daphne==4.0.0
celery==5.3.4
django-celery-beat==2.5.0
Pillow==10.1.0
python-decouple==3.8
dj-database-url==2.1.0
